+++ retroauth-core/src/test/java/com/andretietz/retroauth/stub/TestProvider.java
@@ -1,4 +1,4 @@
-package com.andretietz.retroauth.dummy;
+package com.andretietz.retroauth.stub;
 
 import com.andretietz.retroauth.Provider;
 import com.andretietz.retroauth.TokenStorage;