+++ retroauth-core/src/test/java/com/andretietz/retroauth/stub/TestTokenStorage.java
@@ -1,4 +1,4 @@
-package com.andretietz.retroauth.dummy;
+package com.andretietz.retroauth.stub;
 
 import com.andretietz.retroauth.AuthenticationCanceledException;
 import com.andretietz.retroauth.TokenStorage;
@@ -6,15 +6,18 @@ import com.andretietz.retroauth.TokenStorage;
 /**
  * Created by andre on 02.05.2016.
  */
-public class DummyTokenStorage implements TokenStorage<String, String, String> {
+public class TestTokenStorage implements TokenStorage<String, String, String> {
+
+    public static final String TEST_TOKEN = "token";
+
     @Override
     public String createType(String[] annotationValues) {
         return null;
     }
 
     @Override
-    public String getToken(String s, String s2) throws AuthenticationCanceledException {
-        return null;
+    public String getToken(String owner, String tokenType) throws AuthenticationCanceledException {
+        return TEST_TOKEN;
     }
 
     @Override